% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-population-simulation.R
\name{demo_population_simulation}
\alias{demo_population_simulation}
\title{run several population simulations}
\usage{
demo_population_simulation(nrun, ngeneration, population, fitness, title)
}
\arguments{
\item{nrun}{integer scalar. Number of simulation runs.}

\item{ngeneration}{integer scalar. Number of generations}

\item{population}{integer vector. How many dominant homozygotes, heterozygotes, and recessive homozygotes do the initial population contain. For example, \code{c(2, 5, 3)} means 2 AA, 5 Aa and 3 aa.}

\item{fitness}{numeric vector. The fitness value of AA, Aa and aa. \code{c(1, 1, 1)} means no natural selection.}

\item{title}{string. Plot title.}
}
\value{
an ggplot object
}
\description{
run several population simulations and visualize the result, refer to \code{vignette('pgmcs')} for examples
}
\details{
the underlying workhorse is \code{\link[=population_simulation]{population_simulation()}}
}
\section{Shiny app}{

\code{shiny::runApp(system.file('shiny', package = 'pgs'))}
}

