% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-simulation.R
\name{population_simulation}
\alias{population_simulation}
\title{simulate genetic drift in a finite population}
\usage{
population_simulation(ngeneration, population, fitness = c(1, 1, 1))
}
\arguments{
\item{ngeneration}{integer scalar. Number of generations}

\item{population}{integer vector. How many dominant homozygotes, heterozygotes, and recessive homozygotes do the initial population contain. For example, \code{c(2, 5, 3)} means 2 AA, 5 Aa and 3 aa.}

\item{fitness}{numeric vector. The fitness value of AA, Aa and aa. \code{c(1, 1, 1)} means no natural selection.}
}
\value{
numeric vector. The frequency of allele A.
}
\description{
simulate genetic drift in a finite population for one gene with two alleles
}
\examples{
population_simulation(30, c(0, 10, 0))

}
